# backend/dancee_api/specs/combined.openapi.yaml
openapi: 3.0.0
info:
  title: Dancee API - Complete Documentation
  version: 1.0.0
  description: |
    Complete API documentation for all Dancee backend services.
    
    This unified documentation includes:
    - **Events API**: Event management and user favorites
    - **Scraper API**: Facebook event data extraction
    
    All endpoints are organized by service tags for easy navigation.
  contact:
    name: Dancee API Support
    email: support@dancee.app
servers:
  - url: https://dancee-events.fly.dev
    description: Events API - Production
  - url: http://localhost:8080
    description: Events API - Development
  - url: https://dancee-scraper.fly.dev
    description: Scraper API - Production
  - url: http://localhost:3002
    description: Scraper API - Development
...
paths:
...
  /api/scraper/event:
    get:
      tags:
        - Scraper - Facebook Scraping
      summary: Scrape a single Facebook event
      description: |
        Scrape detailed information about a single Facebook event by providing
        the full Facebook event URL as a query parameter.
        
        **Supported URL formats:**
        - `https://www.facebook.com/events/1987385505448084`
        - `https://facebook.com/events/1987385505448084`
      operationId: scrapeEvent
      servers:
        - url: https://dancee-scraper.fly.dev
        - url: http://localhost:3002
      parameters:
        - name: url
          in: query
          description: Full Facebook event URL
          required: true
          schema:
            type: string
            format: uri
          example: "https://www.facebook.com/events/1987385505448084"
      responses:
        '200':
          description: Event data scraped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacebookEvent'
        '400':
          description: Invalid event ID or URL format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScraperError'
        '404':
          description: Event not found or not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScraperError'
  /api/scraper/events:
    get:
      tags:
        - Scraper - Facebook Scraping
      summary: Scrape a list of events from a Facebook page
      description: |
        Scrape a list of events from a Facebook page, group, or profile.
        
        **Supported formats for pageId:**
        - Page ID: `123456789`
        - Full URL: `https://www.facebook.com/yourpage`
        - Username: `yourpage`
      operationId: scrapeEventList
      servers:
        - url: https://dancee-scraper.fly.dev
        - url: http://localhost:3002
      parameters:
        - name: pageId
          in: query
          description: Facebook page, group, or profile ID or URL
          required: true
          schema:
            type: string
          example: "123456789"
        - name: eventType
          in: query
          description: Filter events by type (upcoming or past)
          required: false
          schema:
            type: string
            enum:
              - upcoming
              - past
          example: upcoming
      responses:
        '200':
          description: Event list scraped successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FacebookEvent'
        '400':
          description: Missing or invalid pageId parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScraperError'
components:
  schemas:
...
    FacebookEvent:
      type: object
      required:
        - id
        - name
        - url
      properties:
        id:
          type: string
          description: Facebook event ID
          example: "1987385505448084"
        name:
          type: string
          description: Event name/title
          example: Prague Salsa Night 2024
        description:
          type: string
          description: Event description
          example: Join us for an amazing night of Salsa and Bachata dancing!
        startTimestamp:
          type: integer
          format: int64
          description: Event start time as Unix timestamp
          example: 1708027200
        endTimestamp:
          type: integer
          format: int64
          description: Event end time as Unix timestamp
          example: 1708048800
        location:
          $ref: '#/components/schemas/EventLocation'
        photo:
          $ref: '#/components/schemas/EventPhoto'
        url:
          type: string
          format: uri
          description: Facebook event URL
          example: https://www.facebook.com/events/1987385505448084
        isOnline:
          type: boolean
          description: Whether the event is online only
          example: false
    EventLocation:
      type: object
      properties:
        name:
          type: string
          example: Lucerna Music Bar
        address:
          type: string
          example: Vodičkova 36, 110 00 Prague, Czech Republic
        city:
          type: string
          example: Prague
        country:
          type: string
          example: Czech Republic
        latitude:
          type: number
          format: double
          example: 50.0813
        longitude:
          type: number
          format: double
          example: 14.4258
    EventPhoto:
      type: object
      properties:
        url:
          type: string
          format: uri
          example: https://scontent.fprg2-1.fna.fbcdn.net/v/t39.30808-6/...
        id:
          type: string
          example: "123456789"
    ScraperError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid event ID or URL format